+++ CHANGELOG.md
@@ -4,6 +4,14 @@ Post 1.0.0 release, the changelog format is based on [Keep a Changelog](https://
 
 ## [Unreleased]
 
+## [1.1.23] (2023-06-27)
+
+- nym-cli: add client identity key signing support ([#3576])
+- Don't fully turn off background task when cover traffic is disabled ([#3596])
+
+[#3576]: https://github.com/nymtech/nym/issues/3576
+[#3596]: https://github.com/nymtech/nym/pull/3596
+
 ## [v1.1.22] (2023-06-20)
 
 - CLI tool for querying network-requesters ([#3539])